# openapi.yaml
openapi: 3.1.0

info:
  title: CURSOR API
  description: |
    API для управления пользователями.
    Реализации: Python (Flask) — api.py, Go — server-go/
  version: 1.0.0
  contact:
    name: CURSOR Project

servers:
  - url: http://localhost:8080
    description: Локальный сервер

paths:
  /adduser:
    post:
      summary: Добавить пользователя
      description: Создаёт нового пользователя в БД. Имя обязательно.
      operationId: addUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddUserRequest'
      responses:
        '201':
          description: Пользователь создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusOk'
        '400':
          description: Невалидный запрос (пустое имя или неверный JSON)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /user/{id}:
    get:
      summary: Получить пользователя по ID
      description: Возвращает данные пользователя. 404 если не найден.
      operationId: getUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
          description: ID пользователя
      responses:
        '200':
          description: Пользователь найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Невалидный ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /activate/{uid}:
    get:
      summary: Активировать uid
      description: |
        Добавляет uid в список активных в фоновом потоке.
        Ответ возвращается сразу, обновление списка происходит асинхронно (~500 ms).
      operationId: activate
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
            minLength: 1
          description: Идентификатор для добавления в активные
      responses:
        '202':
          description: Запрос принят, обработка в фоне
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivateResponse'
        '400':
          description: uid не указан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /slow:
    get:
      summary: Тяжёлые вычисления
      description: Выполняет CPU-интенсивные вычисления. Ответ — число в виде строки.
      operationId: slow
      responses:
        '200':
          description: Результат вычислений
          content:
            text/plain:
              schema:
                type: string
                example: "499999500000000"

  /wrong:
    get:
      summary: Пример обработки ошибок
      description: |
        Эндпоинт, имитирующий внутреннюю ошибку (деление на ноль).
        Демонстрирует корректную обработку panic/исключений.
      operationId: wrong
      responses:
        '200':
          description: В теории — успешный ответ (в текущей реализации не достигается)
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: ok
                  data:
                    type: integer
                    example: 0
        '500':
          description: Ошибка (ожидаемое поведение)
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: error
                  error:
                    type: string
                    description: Текст ошибки

components:
  schemas:
    User:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
          description: ID пользователя
        name:
          type: string
          description: Имя пользователя

    AddUserRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          description: Имя пользователя (непустая строка)

    StatusOk:
      type: object
      properties:
        status:
          type: string
          example: ok

    ActivateResponse:
      type: object
      properties:
        status:
          type: string
          example: processing
        active:
          type: array
          items:
            type: string
          description: Текущий список активных uid

    Error:
      type: object
      properties:
        error:
          type: string
          description: Сообщение об ошибке